Extract-admin-dashboard-data-selectors
